+++ packages/patrol_cli/lib/src/commands/build_ios.dart
@@ -104,6 +104,8 @@ class BuildIOSCommand extends PatrolCommand {
       'PATROL_IOS_APP_NAME': config.ios.appName,
       'PATROL_TEST_LABEL_ENABLED': displayLabel.toString(),
       'INTEGRATION_TEST_SHOULD_REPORT_RESULTS_TO_NATIVE': 'false',
+      'PATROL_TEST_SERVER_PORT': super.testServerPort.toString(),
+      'PATROL_APP_SERVER_PORT': super.appServerPort.toString(),
     }.withNullsRemoved();
 
     final dartDefines = {...customDartDefines, ...internalDartDefines};
@@ -132,6 +134,8 @@ class BuildIOSCommand extends PatrolCommand {
       scheme: flutterOpts.buildMode.createScheme(flavor),
       configuration: flutterOpts.buildMode.createConfiguration(flavor),
       simulator: boolArg('simulator'),
+      appServerPort: super.appServerPort,
+      testServerPort: super.appServerPort,
     );
 
     try {