+++ packages/patrol_cli/lib/src/commands/develop.dart
@@ -160,6 +160,8 @@ class DevelopCommand extends PatrolCommand {
         'PATROL_HOT_RESTART': 'true',
         'PATROL_IOS_INSTALLED_APPS': iOSInstalledAppsEnvVariable,
       },
+      'PATROL_TEST_SERVER_PORT': super.testServerPort.toString(),
+      'PATROL_APP_SERVER_PORT': super.appServerPort.toString(),
     }.withNullsRemoved();
 
     final dartDefines = {...customDartDefines, ...internalDartDefines};
@@ -194,6 +196,8 @@ class DevelopCommand extends PatrolCommand {
       scheme: buildMode.createScheme(iosFlavor),
       configuration: buildMode.createConfiguration(iosFlavor),
       simulator: !device.real,
+      appServerPort: super.appServerPort,
+      testServerPort: super.appServerPort,
     );
 
     final macosOpts = MacOSAppOptions(