+++ packages/patrol_cli/lib/src/runner/patrol_command_runner.dart
@@ -111,6 +111,7 @@ class PatrolCommandRunner extends CompletionCommandRunner<int> {
 
     final iosTestBackend = IOSTestBackend(
       processManager: _processManager,
+      platform: _platform,
       fs: _fs,
       parentDisposeScope: _disposeScope,
       logger: _logger,